extract native Canvas path geometry into moli-canvas
